Add a script to generate a bash script for starting the language server
